When reading FIT files, map developer fields to their native counterpart #27
A developer field, as recorded by a 3rd party application/device can specify a
native counterpart (e.g. a Running Power field can specify that it is a power
field).  If there is no value for the native field, copy the value from the
developer field into the native field.  This means that (assuming the 3rd
party applications are configured correctly), running power data will be
imported into the power series.

There are several limitations to this, in particular it will not consider
developer fields which have no native counterpart like "Leg Stiffness".  See
comment at the top of fit-file.rkt.